Spread out partitioning objects into three files in a subdirectory
